backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
# 2.0.30 breaks on Python 3.13 (__firstlineno__); use a release that includes the TypingOnly fix.
sqlalchemy[asyncio]==2.0.38
alembic==1.13.1
# Use versions with prebuilt wheels for Python 3.13 on Windows (no MSVC required).
psycopg2-binary==2.9.11
asyncpg==0.31.0
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.9
pydantic-settings==2.2.1
python-dotenv==1.0.1
slowapi==0.1.9
bleach==6.1.0
resend==2.0.0
boto3==1.34.101
pytest==8.2.2
pytest-asyncio==0.23.7
httpx==0.27.0
